requirements.txt
accelerate
colorama
cpm_kernels
fastapi
sentencepiece
sse_starlette
streamlit
transformers_stream_generator
unicorn